Update reading file to non-blocking
With fs.promises.readFile() method it is possible to use async/await and mostly important: it reads a file in a non-blocking asynchronous way and mostly important: it is not blocking the node event loop. This is crucial for production and nodejs application. A better solution would be using streams.

The downside in using fs.readFileSync() method, is that we are reading files in a synchronous fashion, this will block node.js event loop and will block other process, only after this syncronous method is finished,  all the remamining node program will be executed.

This PR is just a dropin replacement for the methods and everything will works the same way as before :)